renaming service file
