+++ apps/platform/src/components/Facets/service/facetsService.ts
